Redesigned master selection dialog
The master select dialog now only shows detected systems and components
for reference, and does not allow them to be selected. Accordingly, the
MasterSelect class only returns an interface, not a 3-tuple.

The MasterSelect class now also requires a list of interfaces. The plan
is that it will handle auto-detection of systems and components itself.